@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "django-onfido"
-version = "0.13.1"
+version = "1.0.b2"
 description = "Django app for integration with Onfido."
 license = "MIT"
 authors = ["YunoJuno <code@yunojuno.com>"]
@@ -21,13 +21,14 @@ classifiers = [
     "Programming Language :: Python :: 3 :: Only",
     "Programming Language :: Python :: 3.7",
     "Programming Language :: Python :: 3.8",
+    "Programming Language :: Python :: 3.9",
 ]
 packages = [
     { include = "onfido" }
 ]
 
 [tool.poetry.dependencies]
-python = "^3.7"
+python = "^3.8"
 django = "^2.2 || ^3.0"
 python-dateutil = "*"
 requests =  "*"
