@@ -2,7 +2,7 @@ from unittest import mock
 
 from django.test import TestCase
 
-from onfido import settings
+from onfido import api, settings
 
 
 class SettingsTests(TestCase):
@@ -11,19 +11,21 @@ class SettingsTests(TestCase):
 
     def test_defaults(self):
         """Confirm the default settings exist."""
-        self.assertEqual(settings.API_ROOT, "https://api.onfido.com/v2/")
-        self.assertEqual(settings.API_KEY, None)
+        self.assertEqual(api.API_ROOT, "https://api.onfido.com/v3/")
         self.assertEqual(settings.LOG_EVENTS, True)
-        self.assertEqual(settings.WEBHOOK_TOKEN, None)
         self.assertEqual(settings.TEST_MODE, False)
+        # These may have been set locally
+        # self.assertEqual(settings.API_KEY, None)
+        # self.assertEqual(settings.WEBHOOK_TOKEN, None)
 
     def test_default_report_scrubber(self):
         """Test the report_scrubber default function."""
-        data = {"foo": "bar", "breakdown": {}, "properties": {}}
+        data = {"id": "123", "foo": "bar", "breakdown": {}, "properties": {}}
         # default function should remove breakdown and properties
         data = settings.DEFAULT_REPORT_SCRUBBER(data)
         self.assertFalse("breakdown" in data)
         self.assertFalse("properties" in data)
+        self.assertTrue("id" in data)
 
     # mock scrubber that does nothing and returns the data unchanged
     @mock.patch("onfido.settings.scrub_report_data", lambda d: d)
@@ -49,7 +51,7 @@ class SettingsTests(TestCase):
             "email": "foobar@example.com",
             "first_name": "Foo",
             "gender": None,
-            "href": "/v2/applicants/1b0f8e99-da6b-4ca5-9580-eed99ff691cd",
+            "href": "/v3/applicants/1b0f8e99-da6b-4ca5-9580-eed99ff691cd",
             "id": "1b0f8e97-dc6b-4ca5-9580-eed99ff691cd",
             "id_numbers": [],
             "last_name": "Bar",
@@ -68,7 +70,7 @@ class SettingsTests(TestCase):
             clean,
             {
                 "created_at": "2017-05-03T13:12:17Z",
-                "href": "/v2/applicants/1b0f8e99-da6b-4ca5-9580-eed99ff691cd",
+                "href": "/v3/applicants/1b0f8e99-da6b-4ca5-9580-eed99ff691cd",
                 "id": "1b0f8e97-dc6b-4ca5-9580-eed99ff691cd",
             },
         )
